added more stuff

+++ code/AStarFringeNode.java
@@ -4,17 +4,17 @@ public class AStarFringeNode {
 	
 	public Node dest;
 	
-	public int costFromStart;
+	public double costFromStart;
 	
-	public int estToGoal;
-
+	public double estToGoal;
+	
 	/**
 	 * @param origin
 	 * @param dest
 	 * @param costFromStart
 	 * @param estToGoal
 	 */
-	public AStarFringeNode(Node origin, Node dest, int costFromStart, int estToGoal) {
+	public AStarFringeNode(Node origin, Node dest, double costFromStart, double estToGoal) {
 		this.origin = origin;
 		this.dest = dest;
 		this.costFromStart = costFromStart;
